requirements.txt
Package              Version
-------------------- ---------
annotated-types      0.7.0
anyio                4.11.0
asgiref              3.10.0
black                25.9.0
certifi              2025.10.5
charset-normalizer   3.4.4
click                8.3.0
distro               1.9.0
Django               5.2.7
h11                  0.16.0
httpcore             1.0.9
httpx                0.28.1
idna                 3.11
isort                7.0.0
jiter                0.11.1
jsonpatch            1.33
jsonpointer          3.0.0
langchain-core       1.0.0
langgraph            1.0.1
langgraph-checkpoint 3.0.0
langgraph-prebuilt   1.0.1
langgraph-sdk        0.2.9
langsmith            0.4.38
markdown-it-py       4.0.0
mdurl                0.1.2
mypy_extensions      1.1.0
openai               2.6.0
orjson               3.11.3
ormsgpack            1.11.0
packaging            25.0
pathspec             0.12.1
pip                  25.1.1
platformdirs         4.5.0
pydantic             2.12.3
pydantic_core        2.41.4
Pygments             2.19.2
python-dotenv        1.1.1
pytokens             0.2.0
PyYAML               6.0.3
requests             2.32.5
requests-toolbelt    1.0.0
rich                 14.2.0
sniffio              1.3.1
sqlparse             0.5.3
tenacity             9.1.2
tqdm                 4.67.1
typing_extensions    4.15.0
typing-inspection    0.4.2
urllib3              2.5.0
xxhash               3.6.0
zstandard            0.25.0
